PHR: Rework general file layout
This moves each subcomponent of the router into its own folder. Much
like the link composition, where each larger links are composed of
smaller links, in a similar format. The file layout of the PHR will also
follow a similar pattern, where subdirectories follow a similar layout.
Processors, Classifiers, and Links and other resources are located in
their respective directories.
