--- examples/pi-home-router/src/classifier/mod.rs
@@ -1,2 +1,0 @@
-pub(crate) mod interface;
-pub(crate) mod ethertype;
