+++ examples/pi-home-router/src/interface/processor/annotation_encap.rs
@@ -32,16 +32,3 @@ impl Processor for InterfaceAnnotationEncap {
         })
     }
 }
-
-/// Removes Interface annotations from a packet
-#[derive(Default)]
-pub(crate) struct InterfaceAnnotationDecap;
-
-impl Processor for InterfaceAnnotationDecap {
-    type Input = InterfaceAnnotated<EthernetFrame>;
-    type Output = EthernetFrame;
-
-    fn process(&mut self, packet: Self::Input) -> Option<Self::Output> {
-        Some(packet.packet)
-    }
-}
