+++ examples/pi-home-router/src/interface/processor/mod.rs
@@ -1,0 +1,5 @@
+mod annotation_decap;
+pub(crate) use self::annotation_decap::InterfaceAnnotationDecap;
+
+mod annotation_encap;
+pub(crate) use self::annotation_encap::InterfaceAnnotationEncap;
