--- examples/pi-home-router/src/link/interface/mod.rs
@@ -1,4 +1,0 @@
-mod interface_dispatch;
-mod interface_collect;
-//mod router_intake;
-//mod router_exhaust;
