--- examples/pi-home-router/src/link/mod.rs
@@ -1,6 +1,0 @@
-mod ipv4;
-mod ipv6;
-mod arp;
-mod tracer;
-mod interface;
-mod ndp;
