+++ examples/pi-home-router/src/main.rs
@@ -1,8 +1,6 @@
 mod types;
-mod classifier;
-mod processor;
-mod link;
 mod router;
+mod interface;
 
 use clap::{App, Arg};
 use router::Router;
