--- examples/pi-home-router/src/processor/mod.rs
@@ -1,1 +1,0 @@
-pub(crate) mod interface;
